Improve error flow
