+++ fp-services/src/routes/utils.ts
@@ -1,0 +1,6 @@
+export function generateNonce(): string {
+  const array = new Uint8Array(16);
+  crypto.getRandomValues(array);
+  // @ts-expect-error - works in practice
+  return btoa(String.fromCharCode.apply(null, array));
+}
